refactor(backend): move multipart request and creation boundary out of routes
routes/mediaAssets.ts was 2,858 lines of which only ~470 were the HTTP layer, while every
other file in routes/ is at most 405 lines. Move the shared request primitives and the
upload-session creation application boundary into mediaAssets/multipart/, dropping the
route file to 1,971 lines with all six Hono handlers untouched.
